Working towards more robust gamemode and leaderboard logic

+++ Assets/Scripts/Entity/Dummy/Dummy.cs
@@ -38,7 +38,7 @@ public class Dummy : Entity, IIdentifiable
     {
         if (!baseEntityInitialized) return default;
 
-        ulong TeamId = identification.FetchTeamId();
+        int TeamId = identification.FetchTeamId();
         string EntityName = identification.FetchEntityName();
 
         float healthPercentage = health.HealthPercentage;
